Extract shared useLaminaAssets hook and AssetPickerGrid component (#19)
Move LaminaAsset type and AssetTypeFilter to types.ts. Extract asset
fetching, filtering, and pagination logic into a reusable useLaminaAssets
hook. Extract the asset grid rendering into an AssetPickerGrid component
with optional onSelect callback for picker mode. Refactor AssetBrowser
in LaminaTool to use the shared hook and component.

Closes #19
