+++ src/index.ts
@@ -5,11 +5,17 @@ export { LaminaImageInput } from './components/LaminaFieldAction.js';
 export { LaminaTool } from './tool/LaminaTool.js';
 export { createRegenerateAction } from './actions/regenerateAction.js';
 export { useLamina, LaminaProvider } from './lib/LaminaContext.js';
+export { useLaminaAssets } from './lib/useLaminaAssets.js';
+export { AssetPickerGrid } from './components/AssetPickerGrid.js';
 export type {
+  AssetTypeFilter,
   GeneratedOutput,
   GenerationState,
   GenerationStatus,
+  LaminaAsset,
   LaminaAssetSourceMeta,
   LaminaOAuthConfig,
   LaminaPluginOptions,
 } from './types.js';
+export type { UseLaminaAssetsOptions, UseLaminaAssetsResult } from './lib/useLaminaAssets.js';
+export type { AssetPickerGridProps } from './components/AssetPickerGrid.js';
